Add a ForwardingLogicalPlan to reduce some boiler plate code
