+++ sql/src/main/java/io/crate/planner/operators/FetchOrEval.java
@@ -42,7 +42,6 @@ import io.crate.expression.symbol.FieldReplacer;
 import io.crate.expression.symbol.Function;
 import io.crate.expression.symbol.InputColumn;
 import io.crate.expression.symbol.RefReplacer;
-import io.crate.expression.symbol.SelectSymbol;
 import io.crate.expression.symbol.Symbol;
 import io.crate.expression.symbol.Symbols;
 import io.crate.metadata.DocReferences;
@@ -107,11 +106,10 @@ import static io.crate.planner.operators.OperatorUtils.getUnusedColumns;
  *                Fetch 500
  * </pre>
  */
-public class FetchOrEval implements LogicalPlan {
+public class FetchOrEval extends ForwardingLogicalPlan {
 
     private final FetchMode fetchMode;
     private final boolean doFetch;
-    final LogicalPlan source;
     private final List<Symbol> outputs;
 
     static LogicalPlan.Builder create(LogicalPlan.Builder sourceBuilder,
@@ -170,14 +168,10 @@ public class FetchOrEval implements LogicalPlan {
     }
 
     public FetchOrEval(LogicalPlan source, List<Symbol> outputs, FetchMode fetchMode, boolean doFetch) {
-        this.source = source;
+        super(source);
         this.outputs = outputs;
         this.fetchMode = fetchMode;
         this.doFetch = doFetch;
-    }
-
-    public LogicalPlan source() {
-        return source;
     }
 
     /**
@@ -258,38 +252,8 @@ public class FetchOrEval implements LogicalPlan {
     }
 
     @Override
-    public Map<Symbol, Symbol> expressionMapping() {
-        return source.expressionMapping();
-    }
-
-    @Override
-    public List<AbstractTableRelation> baseTables() {
-        return source.baseTables();
-    }
-
-    @Override
-    public List<LogicalPlan> sources() {
-        return List.of(source);
-    }
-
-    @Override
     public LogicalPlan replaceSources(List<LogicalPlan> sources) {
         return new FetchOrEval(Lists2.getOnlyElement(sources), outputs, fetchMode, doFetch);
-    }
-
-    @Override
-    public Map<LogicalPlan, SelectSymbol> dependencies() {
-        return source.dependencies();
-    }
-
-    @Override
-    public long numExpectedRows() {
-        return source.numExpectedRows();
-    }
-
-    @Override
-    public long estimatedRowSize() {
-        return source.estimatedRowSize();
     }
 
     private ExecutionPlan planWithFetch(PlannerContext plannerContext,
