+++ sql/src/main/java/io/crate/planner/operators/Filter.java
@@ -24,13 +24,11 @@ package io.crate.planner.operators;
 
 import io.crate.analyze.OrderBy;
 import io.crate.analyze.QueryClause;
-import io.crate.analyze.relations.AbstractTableRelation;
 import io.crate.collections.Lists2;
 import io.crate.data.Row;
 import io.crate.execution.dsl.projection.FilterProjection;
 import io.crate.execution.dsl.projection.builder.ProjectionBuilder;
 import io.crate.expression.symbol.Literal;
-import io.crate.expression.symbol.SelectSymbol;
 import io.crate.expression.symbol.Symbol;
 import io.crate.metadata.RowGranularity;
 import io.crate.planner.ExecutionPlan;
@@ -40,15 +38,13 @@ import io.crate.types.DataTypes;
 import javax.annotation.Nullable;
 import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
 import static io.crate.planner.operators.LogicalPlanner.extractColumns;
 
-public final class Filter implements LogicalPlan {
+public final class Filter extends ForwardingLogicalPlan {
 
     final Symbol query;
-    final LogicalPlan source;
 
     static LogicalPlan.Builder create(LogicalPlan.Builder sourceBuilder, @Nullable QueryClause queryClause) {
         if (queryClause == null) {
@@ -81,16 +77,12 @@ public final class Filter implements LogicalPlan {
     }
 
     public Filter(LogicalPlan source, Symbol query) {
-        this.source = source;
+        super(source);
         this.query = query;
     }
 
     public Symbol query() {
         return query;
-    }
-
-    public LogicalPlan source() {
-        return source;
     }
 
     @Override
@@ -114,43 +106,8 @@ public final class Filter implements LogicalPlan {
     }
 
     @Override
-    public List<Symbol> outputs() {
-        return source.outputs();
-    }
-
-    @Override
-    public Map<Symbol, Symbol> expressionMapping() {
-        return source.expressionMapping();
-    }
-
-    @Override
-    public List<AbstractTableRelation> baseTables() {
-        return source.baseTables();
-    }
-
-    @Override
-    public List<LogicalPlan> sources() {
-        return List.of(source);
-    }
-
-    @Override
     public LogicalPlan replaceSources(List<LogicalPlan> sources) {
         return new Filter(Lists2.getOnlyElement(sources), query);
-    }
-
-    @Override
-    public Map<LogicalPlan, SelectSymbol> dependencies() {
-        return source.dependencies();
-    }
-
-    @Override
-    public long numExpectedRows() {
-        return source.numExpectedRows();
-    }
-
-    @Override
-    public long estimatedRowSize() {
-        return source.estimatedRowSize();
     }
 
     @Override
