+++ sql/src/main/java/io/crate/planner/operators/GroupHashAggregate.java
@@ -23,7 +23,6 @@
 package io.crate.planner.operators;
 
 import io.crate.analyze.OrderBy;
-import io.crate.analyze.relations.AbstractTableRelation;
 import io.crate.collections.Lists2;
 import io.crate.data.Row;
 import io.crate.execution.dsl.phases.ExecutionPhases;
@@ -34,7 +33,6 @@ import io.crate.execution.dsl.projection.builder.ProjectionBuilder;
 import io.crate.execution.engine.pipeline.TopN;
 import io.crate.expression.symbol.AggregateMode;
 import io.crate.expression.symbol.Function;
-import io.crate.expression.symbol.SelectSymbol;
 import io.crate.expression.symbol.Symbol;
 import io.crate.metadata.RowGranularity;
 import io.crate.metadata.doc.DocTableInfo;
@@ -50,17 +48,15 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Map;
 
 import static io.crate.planner.operators.LogicalPlanner.NO_LIMIT;
 import static io.crate.planner.operators.LogicalPlanner.extractColumns;
 
-public class GroupHashAggregate implements LogicalPlan {
+public class GroupHashAggregate extends ForwardingLogicalPlan {
 
     private static final String DISTRIBUTED_MERGE_PHASE_NAME = "distributed merge";
     final List<Function> aggregates;
     final List<Symbol> groupKeys;
-    final LogicalPlan source;
     private final List<Symbol> outputs;
 
     public static Builder create(Builder source, List<Symbol> groupKeys, List<Function> aggregates) {
@@ -73,7 +69,7 @@ public class GroupHashAggregate implements LogicalPlan {
     }
 
     GroupHashAggregate(LogicalPlan source, List<Symbol> groupKeys, List<Function> aggregates) {
-        this.source = source;
+        super(source);
         this.outputs = Lists2.concat(groupKeys, aggregates);
         this.groupKeys = groupKeys;
         this.aggregates = aggregates;
@@ -152,28 +148,8 @@ public class GroupHashAggregate implements LogicalPlan {
     }
 
     @Override
-    public Map<Symbol, Symbol> expressionMapping() {
-        return source.expressionMapping();
-    }
-
-    @Override
-    public List<AbstractTableRelation> baseTables() {
-        return source.baseTables();
-    }
-
-    @Override
-    public List<LogicalPlan> sources() {
-        return List.of(source);
-    }
-
-    @Override
     public LogicalPlan replaceSources(List<LogicalPlan> sources) {
         return new GroupHashAggregate(Lists2.getOnlyElement(sources), groupKeys, aggregates);
-    }
-
-    @Override
-    public Map<LogicalPlan, SelectSymbol> dependencies() {
-        return source.dependencies();
     }
 
     private ExecutionPlan createMerge(PlannerContext plannerContext,
@@ -213,17 +189,6 @@ public class GroupHashAggregate implements LogicalPlan {
                    ((DocTableInfo) ((Collect) source).tableInfo),
                    ((Collect) source).where,
                    groupKeys);
-    }
-
-    @Override
-    public long numExpectedRows() {
-        // We don't have any cardinality estimates
-        return source.numExpectedRows();
-    }
-
-    @Override
-    public long estimatedRowSize() {
-        return source.estimatedRowSize();
     }
 
     @Override
