+++ sql/src/main/java/io/crate/planner/operators/MultiPhase.java
@@ -23,13 +23,11 @@
 package io.crate.planner.operators;
 
 import io.crate.analyze.OrderBy;
-import io.crate.analyze.relations.AbstractTableRelation;
 import io.crate.analyze.relations.AnalyzedRelation;
 import io.crate.collections.Lists2;
 import io.crate.data.Row;
 import io.crate.execution.dsl.projection.builder.ProjectionBuilder;
 import io.crate.expression.symbol.SelectSymbol;
-import io.crate.expression.symbol.Symbol;
 import io.crate.planner.ExecutionPlan;
 import io.crate.planner.MultiPhasePlan;
 import io.crate.planner.PlannerContext;
@@ -45,9 +43,8 @@ import java.util.Map;
  * This is the {@link LogicalPlan} equivalent of the {@link MultiPhasePlan} plan.
  * It's used to describe that other logical plans needed to be executed first.
   */
-public class MultiPhase implements LogicalPlan {
+public class MultiPhase extends ForwardingLogicalPlan {
 
-    final LogicalPlan source;
     private final Map<LogicalPlan, SelectSymbol> subQueries;
 
     public static LogicalPlan createIfNeeded(LogicalPlan source,
@@ -61,7 +58,7 @@ public class MultiPhase implements LogicalPlan {
     }
 
     private MultiPhase(LogicalPlan source, Map<LogicalPlan, SelectSymbol> subQueries) {
-        this.source = source;
+        super(source);
         HashMap<LogicalPlan, SelectSymbol> allSubQueries = new HashMap<>(source.dependencies());
         allSubQueries.putAll(subQueries);
         this.subQueries = Collections.unmodifiableMap(allSubQueries);
@@ -81,26 +78,6 @@ public class MultiPhase implements LogicalPlan {
     }
 
     @Override
-    public List<Symbol> outputs() {
-        return source.outputs();
-    }
-
-    @Override
-    public Map<Symbol, Symbol> expressionMapping() {
-        return source.expressionMapping();
-    }
-
-    @Override
-    public List<AbstractTableRelation> baseTables() {
-        return source.baseTables();
-    }
-
-    @Override
-    public List<LogicalPlan> sources() {
-        return List.of(source);
-    }
-
-    @Override
     public LogicalPlan replaceSources(List<LogicalPlan> sources) {
         return new MultiPhase(Lists2.getOnlyElement(sources), subQueries);
     }
@@ -108,16 +85,6 @@ public class MultiPhase implements LogicalPlan {
     @Override
     public Map<LogicalPlan, SelectSymbol> dependencies() {
         return subQueries;
-    }
-
-    @Override
-    public long numExpectedRows() {
-        return source.numExpectedRows();
-    }
-
-    @Override
-    public long estimatedRowSize() {
-        return source.estimatedRowSize();
     }
 
     @Override
